Added some tests for incremental search backward and forward. These tests don't pass all, so it's work in progress.
Also achieved it, that several SWTBot tests can be executed in a row.

Also added tests for the MatchEvaluatorFromItsFunctionBodyGenereator
checking whether is produces the correct exception with the correct
message in case of an unsuccessful compile.

And tested the RegexUtils class again, where a Pattern.CASE_INSENSITIVE
flag is set, and where a match evaluator throws an exception.