Added log4j, so you can inspect errors if the plugin is run from a deployed instance.
Added a test to FindReplaceDialogTest for the case when the matches are
just positions (so matches of length zero). These can't be found by
IFindReplaceTarget, so I reported a bug here
https://bugs.eclipse.org/bugs/show_bug.cgi?id=499327.

Added a test for the case when there is a lookahead that captures a
group that is referred to by the match evaluator. (Doesn't work for
normal regex replace either)

Added a test for the case when there is a lookbehind that captures a
group that is referred to by the match evaluator. (Doesn't work for
normal regex replace either)

Added all these tests also for RegexUtils.replaceAll method. Here they
succeed.

Added the help icon to FindReplaceDialog again.