Set up log4j in the SWTBot tests for the FindReplaceDialog, too.
Added flags: in FindReplaceDialogTests. Added the @Ignore annotation to
the 2 tests, that treat the case where the matches are just positions. I
will not be able to fix this, nor is it my responsibility.
The tests that check whether a group in the lookbehind or lookahead get
caught work now with the changes made to the other files.

Added Pattern.MULTILINE to the FindReplaceDialog and to my fake
FindReplaceTarget!

Added tests for the function RegexUtils.getReplaceStringOfFirstMatch
that check whether it can find a match starting from an offset where the
lookbehind comes before that offset. And it can do this.

log4j is now configured in the Activator class. It creates a logfile in
the same directory as eclipse.exe.