Migrated the plugin to use Eclipse 4.
So I got rid of the dependency to the compatibility layer.
I removed everything from the plugin.xml file except for the extension
org.eclipse.e4.workbench.model.
Instead I created fragment.e4xmi which defines a menu contribution to
the main menu with an attached handler. This handler works as the
previous one, but uses dependency injection.