Removed org.eclipse.e4.core.contexts and org.eclipse.e4.core.services from the required bundles in MANIFEST.mf.
WIP: Added a test to check whether all dialog settings are stored
between popups of the dialog.

Removed some more of the hardcoded values and put them in
DialogSettingsConstants!! This is not all tested yet!

Changed FindReplaceDialog so that is works also with Scope "Selected
lines". This is however only tested manually. The old tests still pass
fortunately.