The JavacLocator is now injected into the FindReplaceDialog as a dependency.
Added a test, that the dialog still works when the JavacLocator( which
is faked in the test) can't find javac.exe. The test also checks that
"NoJavaCompilerSetException occured" is displayed in the status label
and the status line.

Added a test to see whether it works to change the match evaluator code
and to see whether the new one is compiled and used.

Added a test to see that the match evaluator code isn't saved to the
dialog settings if this match evaluator can't be compiled.