Added Pattern.MULTILINE to the flags label in the dialog.
Added a test that checks whether the replacement with a match evaluator
also works when the MULTILINE mode matters in the search.

Changed new FileDialog( Display.getCurrent().getActiveShell ) to new
FileDialog ( getShell() ) and tested it.

!!! Fixed bug when the JavacLocator can't find javac. Previously it
returned null and then a null pointer occured. This was of course a
serious mistake.