The FindReplaceDialogTest class now uses an ordinary org.eclipse.jface.text.Document to modify. The initial window in these ui tests is now larger and has an appropriate title.
You can now also make manual tests with a standalone SWT application
residing in manualtests/MainWindow.java.

The match evaluator field now got a heightHint of 75, which looks like a
decent preferred height.
Also the shell is now always packed if something becomes invisible or is
excluded from the layout process. The flags label is now originally
visible. This was a bug! It was solved by adding the condition

if(getShell().isVisible() )
forceRecalculationOfLayoutAndARepaint();

to above lines of code.
Before the text of the flags label wasn't yet set!