Removed log4j configuration from the automated ui tests.
Added two tests. These test whether a named group can be referenced from
the MatchEvaluator. The second test doesn't pass because the regular
expression (?![A-Z]+\s*=)([\w_]+)(\s*=) can't be found in e.g.
public static final int caseInsensitive = 0; This seems to be another
Eclipse bug.

Now the interface IMatchEvaluator is changed so that the method
evaluateMatch receives a Matcher as an argument. That's why the Matcher
class has a method String group(String groupName) and the MatchResult
class doesn't have this method.
In the GUI, there is still MatchResult displayed, but that's fine
because Matcher implements MatchResult.