Added a test, where the user enters not only one member function for his MatchEvaluator but also fields and other instance methods.
In order to make above test pass, I changed the way the user's code is
compiled and loaded by a class loader. This process now requires Java
1.8 because the interface java.util.function.Function<T,R> is used.

There will be no need for the future to provide the possibility for the
user to add additional imports. He can use the fully qualified class
names instead.